add other method for wire up
