+++ Source/Our.Umbraco.RedirectsViewer/Our.Umbraco.RedirectsViewer.csproj
@@ -269,13 +269,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Compositions\UmbracoStartup.cs" />
     <Compile Include="Controllers\RedirectsApiController.cs" />
     <Compile Include="Controllers\UserGroupsApiController.cs" />
-    <Compile Include="Mapping\UsergroupMapping.cs" />
+    <Compile Include="Mapping\UserGroupProfile.cs" />
     <Compile Include="Models\RedirectSave.cs" />
     <Compile Include="Models\UserGroupDisplay.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UmbracoStartup.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
